--- a/pilot/pkg/proxy/envoy/v2/pushqueue.go
+++ b/pilot/pkg/proxy/envoy/v2/pushqueue.go
@@ -63,17 +63,23 @@ func (p *PushQueue) Enqueue(proxy *XdsConnection, pushInfo *PushEvent) {
 	} else {
 		event.push = pushInfo.push
 		event.full = event.full || pushInfo.full
-
-		edsUpdates := map[string]struct{}{}
-		for endpoint := range pushInfo.edsUpdatedServices {
-			edsUpdates[endpoint] = struct{}{}
+		// When full push, do not care about edsUpdatedServices
+		if !event.full {
+			edsUpdates := map[string]struct{}{}
+			for endpoint := range pushInfo.edsUpdatedServices {
+				edsUpdates[endpoint] = struct{}{}
+			}
+			for endpoint := range event.edsUpdatedServices {
+				edsUpdates[endpoint] = struct{}{}
+			}
+			event.edsUpdatedServices = edsUpdates
+		} else {
+			event.edsUpdatedServices = nil
 		}
-		for endpoint := range event.edsUpdatedServices {
-			edsUpdates[endpoint] = struct{}{}
-		}
-		event.edsUpdatedServices = edsUpdates
 	}
-	p.cond.Signal()
+	if len(p.connections) > 0 {
+		p.cond.Signal()
+	}
 }
 
 // Remove a proxy from the queue. If there are no proxies ready to be removed, this will block