@istio-testing 2019-08-07
@hzxuzhonghu: The following tests **failed**, say `/retest` to rerun them all:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
e2e-simpleTests-master | 0cf9b7bb8c3ebd84cd0b145b6c99ec674c62bfc5 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/16097/e2e-simpleTests-master/5131) | `/test e2e-simpleTests-master`
e2e-simpleTests-distroless-master | 0cf9b7bb8c3ebd84cd0b145b6c99ec674c62bfc5 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/16097/e2e-simpleTests-distroless-master/872) | `/test e2e-simpleTests-distroless-master`
e2e-simpleTestsMinProfile-master | 0cf9b7bb8c3ebd84cd0b145b6c99ec674c62bfc5 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/16097/e2e-simpleTestsMinProfile-master/2469) | `/test e2e-simpleTestsMinProfile-master`
istio-unit-tests-master | 0cf9b7bb8c3ebd84cd0b145b6c99ec674c62bfc5 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/16097/istio-unit-tests-master/5040) | `/test istio-unit-tests-master`



<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->