@hzxuzhonghu 2019-08-08
```

func BenchmarkEnqueue(b *testing.B) {
	con := &XdsConnection{ConID: "proxy"}
	edsUpdatedServices1 := map[string]struct{}{}
	edsUpdatedServices2 := map[string]struct{}{}

	for i := 0; i < 20; i++ {
		if i < 10 {
			edsUpdatedServices1[fmt.Sprintf("svc-%d", i)] = struct{}{}
		} else {
			edsUpdatedServices2[fmt.Sprintf("svc-%d", i)] = struct{}{}

		}
	}

	for i := 0; i < b.N; i++ {
		p := NewPushQueue()
		p.Enqueue(con, &PushEvent{full:true, edsUpdatedServices: edsUpdatedServices1})
		p.Enqueue(con, &PushEvent{full:true, edsUpdatedServices: edsUpdatedServices2})

	}
}
```


Without this pr: 
```
goos: linux
goarch: amd64
pkg: istio.io/istio/pilot/pkg/proxy/envoy/v2
BenchmarkEnqueue-32    	  200000	      6477 ns/op
PASS
ok  	istio.io/istio/pilot/pkg/proxy/envoy/v2	3.476s
```

With this:

```
goos: linux
goarch: amd64
pkg: istio.io/istio/pilot/pkg/proxy/envoy/v2
BenchmarkEnqueue-32    	 1000000	      1233 ns/op
PASS
ok  	istio.io/istio/pilot/pkg/proxy/envoy/v2	3.361s
```